Added new configuration tests Separated sensor readings and display test into it's own test section Fixed & Enhanced readings tests Removed manually created text tests due to overhead on maintaining them Created tests folder for program tests Moved test_http_server.py into it Created separate file for test configurations
